:art:rename: post 페이지 상수 파일명 변경
